Safari-611.1.10/LayoutTests/crypto/subtle/hkdf-import-key-derive-bits-expected.txt

Test HKDF deriveBits operation with an imported raw key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(derivedKey) is expectedDerivedKey
PASS successfullyParsed is true

TEST COMPLETE
